RDKB-57845: cci providing multiple Beacons even when there is no config changes
Reason for change: cci should not create the beacons if there is no change in the vap subdocuments.
Test Procedure: 1) apply the same vap configuration multiple times, there should not be any beacons created. Make sure there should not be any Forceapply present in the subdocument with this test.
2) Improved logging in error prints in onewifi logs
Risks: Low
Priority: P1

Change-Id: If37153495a8cff79e22f46e9218d25d42cae5ea6
Signed-off-by: Ramakanth Taraka <ramakanth.taraka@sky.uk>