RDKB-60023: Failed to Set Default country code as `IE` for UK patner (#377)
Reason for change: As part of XER10 ROI, as using same partner as
sky-uk, when we set default country code as IE (Ireland), onewifi
failed to set it. This occurred when we change default RegionCode
in partners_defaults.json as IE for XER10 under sky-uk partner.

The issue is due to failure in copying default country code as
IE in wifi_db_apis.c.

Here corrected the string copy opertauon to properly get the
country code from bootstrap.json.

Test Procedure:
1. Load the image
2. Make sure default country code came as IE instead of GB for XER10 UK.

Priority:P1

Signed-off-by: vysakhav <vysakhav@protonmail.com>