RDKB-60494: EM translator: Update vap_mode in em_bss_info_t
Reason for change: The vap_mode needs to be updated to the
controller for building the right topology. Ensure that the
vap_mode is correctly updated in em object in agent which in
turn is sent to controller as part of topology response.

Test Procedure: Tested multiple ssid change and channel change across
colocated agent and remote agent scenario.

Risks: Low
Priority: P1