RDKBWIFI-48: Pass vap password through to easymesh
Reason for change: With this, unified-wifi-mesh can read the SSID and
password of the backhaul mesh, then tell the backhaul sta that it should
connect to that network. `unified-wifi-mesh` should be in charge of
telling OneWifi what it should associate to. `OneWifi` shouldn't blindly
associate to a network.

Test Procedure: Ensure the `mesh_ap_passphrase` variable is populated in
the `em_bss_t` struct.

Risks: Very Low

Priority: P1

Signed-off-by: Benjamin Carlson <b.carlson@cablelabs.com>